Remove module test code from cards_management.c to test folder.
